feat: CD script support ability to skip login

+++ lib/templates/cd.sh
@@ -15,22 +15,25 @@ DOCKER_BUILD_CMD="<%=continuous_deployment.docker_build_cmd || 'build -f Dockerf
 CI_COMMIT_SHA=$(git rev-parse --verify HEAD)
 DEPLOY_NAME="${IMAGE_NAME}:${CI_COMMIT_SHA}"
 LATEST_NAME="${IMAGE_NAME}:latest"
-AUTH_PATH="$SCRIPT_DIR/k8s-auth-token.json"
 
-rm -f -- $AUTH_PATH
-echo $KB_AUTH_TOKEN >> $AUTH_PATH
+if [ ! -z "$KB_AUTH_TOKEN" ]
+then
+  AUTH_PATH="$SCRIPT_DIR/k8s-auth-token.json"
+  rm -f -- $AUTH_PATH
+  echo $KB_AUTH_TOKEN >> $AUTH_PATH
 
-## ***** GOOGLE CONNECTOR
-# Download and install Google Cloud SDK
-if [ -z "$(which gcloud)" ]; then
-  export CLOUDSDK_CORE_DISABLE_PROMPTS=1; curl https://sdk.cloud.google.com | bash && source /home/runner/google-cloud-sdk/path.bash.inc &&  gcloud --quiet components update kubectl
+  ## ***** GOOGLE CONNECTOR
+  # Download and install Google Cloud SDK
+  if [ -z "$(which gcloud)" ]; then
+    export CLOUDSDK_CORE_DISABLE_PROMPTS=1; curl https://sdk.cloud.google.com | bash && source /home/runner/google-cloud-sdk/path.bash.inc &&  gcloud --quiet components update kubectl
+  fi
+
+  # Connect to cluster
+  gcloud auth activate-service-account --key-file $AUTH_PATH --project $PROJECT_NAME
+  gcloud docker --authorize-only --project $PROJECT_NAME
+  gcloud container clusters get-credentials $CLUSTER_NAME --region $CLUSTER_REGION
+  ## ***** END GOOGLE CONNECTOR
 fi
-
-# Connect to cluster
-gcloud auth activate-service-account --key-file $AUTH_PATH --project $PROJECT_NAME
-gcloud docker --authorize-only --project $PROJECT_NAME
-gcloud container clusters get-credentials $CLUSTER_NAME --region $CLUSTER_REGION
-## ***** END GOOGLE CONNECTOR
 
 
 ## Build and push containers
